+++ src/test/java/com/solace/spring/cloud/stream/binder/springBootTests/healthindicator/brokenconfig/BrokenConfigApp.java
@@ -1,0 +1,22 @@
+package com.solace.spring.cloud.stream.binder.springBootTests.healthindicator.brokenconfig;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.web.SecurityFilterChain;
+
+@SpringBootApplication
+public class BrokenConfigApp {
+    public static void main(String[] args) {
+        SpringApplication.run(BrokenConfigApp.class, args);
+    }
+
+    @Bean
+    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+        http.authorizeHttpRequests(requests -> requests
+                .requestMatchers("/actuator/**").permitAll()
+                .anyRequest().authenticated());
+        return http.build();
+    }
+}